refactor: split prefetch plan handling into lineairdb_prefetch
Move the @_prefetch_plan DSL parsing and execution helpers out of ha_lineairdb.cc into lineairdb_prefetch.{cc,hh}, exposing thd_can_use_prefetch, thd_has_prefetch_plan, and execute_prefetch_plan_if_present through the new header, and drop the now-unused lineairdb_keyenc.hh include from the handler.
